Remove default ubuntu user
The Ubuntu 24.04 base image now comes with a default user "ubuntu" that
has a uid of 1000. This breaks the ability to use uid 1000 for the
BORG_UID environment variable because this uid is now occupied by the
default user.

Fixes the following log message:

```
usermod: UID '1000' already exists
usermod: no changes
```

Signed-off-by: Felix Kaechele <felix@kaechele.ca>